Optimized JSLexer with lookahead caching
Before, each call to `peek()` would result in a new Token being parsed,
which is clearly inefficient. Now, it stores the previous value when
`peek()` is first called for a new token, which will be returned by
future calls to `peek()` and `nextToken()`. The cache is cleared when
`skip()` or `nextToken()` is called.
Tests passing.