Check invalid lambda expressions better
Because `*=>()` will always fail to parse (Empty parentheses aren't a
valid lambda body), most of the invalid tests wouldn't fail even if the
code being tested was wrong.
Turns out that there *is* a bug that was masked by this, where required
parameters following optional ones doesn't throw an error.